Refactor: eliminate forward declarations by extracting common enums
Create AgentEnums.H to hold shared enum definitions and helper functions used by both AgentDefinitions.H and DiseaseParm.H. This eliminates circular dependency and removes need for forward declarations.

Changes:
- Create src/AgentEnums.H with AgeGroups, SchoolType, IntIdx, IntIdxDisease, RealIdxDisease, Status, DiseaseStats enums and i0(), r0() helper functions
- Update src/AgentDefinitions.H to include AgentEnums.H and remove duplicate definitions
- Update src/DiseaseParm.H to include AgentEnums.H instead of AgentDefinitions.H
- Include DiseaseParm.H in AgentDefinitions.H after basic definitions to provide complete type for setAgentData function

Build completes without warnings or errors.
