+++ src/AgentDefinitions.H
@@ -6,6 +6,7 @@
 #define _AGENT_DEF_H_
 
 #include <AMReX_Particles.H>
+#include "AgentEnums.H"
 
 namespace ExaEpi {
 /*! Maximum number of diseases */
@@ -16,72 +17,6 @@ const int max_num_diseases = 10;
 struct RealIdx {
     enum {
         nattribs = 0 /*!< number of real-type attribute*/
-    };
-};
-
-/*! \brief Disease-specific Real-type Runtime-SoA attributes of agent */
-struct RealIdxDisease {
-    enum {
-        /* Disease counter starts after infection. */
-        treatment_timer = 0, /*!< Timer since hospital admission */
-        disease_counter,     /*!< Counter since start of infection */
-        prob,                /*!< Probability of infection */
-        latent_period,       /*!< Time until infectious, which could be before symptoms appear */
-        infectious_period,   /*!< Length of time infectious */
-        incubation_period,   /*!< Time until symptoms appear */
-        hospital_delay,      /*!< Delay after symptom onset until hospital treatment is sought */
-        hospital_random,     /*!< Random number that decides hospitaliation */
-        nattribs             /*!< number of real-type attribute*/
-    };
-};
-
-/*! \brief Integer-type SoA attributes of agent */
-struct IntIdx {
-    enum {
-        age_group = 0, /*!< Age group (under 5, 5-17, 18-29, 30-64, 65+) */
-        family,        /*!< Family ID */
-        home_i,        /*!< home location index */
-        home_j /*!< home location index */,
-        work_i /*!< work location index */,
-        work_j /*!< work location index */,
-        hosp_i /*!< hosp location index */,
-        hosp_j /*!< hosp location index */,
-        trav_i /*!< air travel location index */,
-        trav_j /*!< air travel location index */,
-        nborhood,      /*!< home neighborhood ID */
-        hh_cluster,    /*!< household cluster ID */
-        school_grade,  /*!< school grade, including universities */
-        school_id,     /*!< ID for a given school */
-        school_closed, /*!< 0 for open, 1 for closed */
-        naics,         /*!< industry NAICS code for business employed at */
-        workgroup,     /*!< workgroup ID */
-        work_nborhood, /*!< work neighborhood ID */
-        withdrawn,     /*!< quarantine status */
-        random_travel, /*!< on long distance travel? */
-        air_travel,    /*!< on long distance travel by Air? */
-        nattribs       /*!< number of integer-type attribute */
-    };
-};
-
-/*! \brief Disease-specific Integer-type Runtime-SoA attributes of agent */
-struct IntIdxDisease {
-    enum {
-        status = 0,  /*!< Disease status (#Status) */
-        symptomatic, /*!< currently symptomatic? 0: no, but will be, 1: yes, 2: no, and will remain so until recovered */
-        nattribs     /*!< number of integer-type attribute */
-    };
-};
-
-/*! \brief School Type  */
-struct SchoolType {
-    enum {
-        none, /*!< Not in school */
-        college,
-        high,
-        middle,
-        elem,
-        daycare,
-        total
     };
 };
 
@@ -97,54 +32,6 @@ struct SchoolCensusIDType {
         total
     };
 };
-
-/*! \brief Age Group */
-struct AgeGroups {
-    enum {
-        u5 = 0,  /*!< Under 5 */
-        a5to17,  /*!< 5-17 */
-        a18to29, /*!< 18-29 */
-        a30to49, /*!< 30-49 */
-        a50to64,
-        o65,     /*!< over 65 */
-        total    /*!< number of age groups */
-    };
-};
-
-/*! \brief Disease status */
-struct Status {
-    enum {
-        never = 0,   /*!< never infected */
-        infected,    /*!< infected */
-        immune,      /*!< no longer infected, immune. lasts 6 months. */
-        susceptible, /*!< no longer infected, no longer immnune */
-        dead         /*!< passed away */
-    };
-};
-
-/*! \brief Disease statistics */
-struct DiseaseStats {
-    enum {
-        hospitalization = 0, /*!< number of current hospitalizations */
-        ICU,                 /*!< number of current ICU cases */
-        ventilator,          /*!< number of current ventilator cases */
-        death,               /*!< number of deaths */
-        new_cases            /*!< number of new infections each day */
-    };
-};
-
-/*! \brief Compute index offsets for runtime int-type disease attributes */
-AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-int i0 (const int a_d /*!< Disease index */) {
-    return a_d * IntIdxDisease::nattribs;
-}
-
-/*! \brief Compute index offsets for runtime real-type disease attributes */
-AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-int r0 (const int a_d /*!< Disease index */) {
-    return a_d * RealIdxDisease::nattribs;
-}
-
 /*! \brief Disease symptom status */
 struct SymptomStatus {
     enum {
@@ -331,6 +218,9 @@ bool isNewlyHospitalized (const int a_idx,      /*!< Agent index */
     int hospital_delay = (int)(amrex::Math::floor(a_ptd.m_runtime_rdata[r0(a_d) + RealIdxDisease::hospital_delay][a_idx]));
     return disease_counter == incb_period + hospital_delay;
 }
+
+// Include DiseaseParm after basic definitions to avoid circular dependency
+#include "DiseaseParm.H"
 
 template <typename PTDType>
 AMREX_GPU_DEVICE AMREX_FORCE_INLINE
